Add files via upload
The objective is to address a hypothetical business problem for a Flipkart Authorized Seller as well as the Mobile Company. According to the problem the individual and the company management is looking to increase their mobile sales. For this, they are looking for the best product, brand, specification and deals that can generate the most revenue with the least amount of investment and budget constraints.